requirements.txt
google-genai==1.45.0

playwright==1.55.0

python-dotenv==1.1.1